add recall-server implementation
Add an implementation for a simple recall/concentration server
build on Node.js.  The only dependency is underscore.

A simple routing system is used to dispatch requests to a
controller.  See server.js for an example route table and Nails.js
for the guts.

The Game class is basically a JavaScript clone of the Python class
recall.game.OfflineGame in https://github.com/frasertweedale/recall.

Games are currently kept in memory forever so the server (and
possibly its host) is vulnerable to DoS through memory exhaustion.